Increase JVM heap space for activemq
Not having this in leads to the following error during startup:

2014-10-10 10:37:52,552 | ERROR | Memory Usage for the Broker (1024 mb) is more than the maximum available for the JVM: 981 mb - resetting to 70% of maximum available: 686 mb | org.apache.activemq.broker.BrokerService | main

The reason for this error is that the JVM is configured to use one GB of heap in the activemq.init script:

# Set jvm memory configuration
if [ -z "$ACTIVEMQ_OPTS_MEMORY" ] ; then
    ACTIVEMQ_OPTS_MEMORY="-Xms1G -Xmx1G"
fi

At the same time, the internal broker of activemq seems to be asking 
for one GB of memory for its systemUsage by default (unless this is 
configured differently in activemq.xml). Naturally, it can't have the entire
heap for itself.

Note that there is a parallel pull request (#1) that configures activemq.xml
explicitly to use 1 GB of memory.